Got the map up!
Entirely re-did the map sprite, used a map layout instead of a sprite,
and had to redo it (add spaces) for my own sanity. The map is on, but
there is no collision at this point in time, that should be up tomorrow
(3/2/14) but I'll have to see.
I also made it so one does not have to spam the arrow keys to move
about. pac-man moves forward on his own.
Food should be next week or so, and in sake of time, I have scrapped the
Missile command idea, there will be a very simple AI enemy.